Cast select arg to silence warning (ignored in winsock according to MS doc)

+++ libvisiontransfer/visiontransfer/imagetransfer.cpp
@@ -770,7 +770,7 @@ bool ImageTransfer::Pimpl::selectSocket(bool read, bool wait) {
         tv.tv_usec = 0;
     }
 
-    if(select(sock+1, (read ? &fds : nullptr), (!read ? &fds : nullptr), nullptr, &tv) <= 0) {
+    if(select(((int)sock)+1, (read ? &fds : nullptr), (!read ? &fds : nullptr), nullptr, &tv) <= 0) {
         // The socket is currently not ready
         return false;
     }
